Hiding overflow for cards
The combination of shadow-box with border-radius that we're using for our cards causes some weird problems when a child of the card has background-color. The most recent I've had was with the lower left border having a little white glitch.

This can be fixed with `overflow: hidden`, but since it's so common (and it took me a whihle to investigate as well) it's best to have that fixed here, so no one else falls into this again.