Fix prop type errors for icons inside the layout sidebar
The icons we use are not the default icons for react semantic ui, so passing them directly as `icon` throws some warnings in the console for icons that were not expected.

We need to instead pass them as `className`. To have this be transparent for people using Layout, I'm doing this conversion inside `Layout.SidebarItem` and `Layout.SidebarSubMenu`.