Add transition for dimmers
The current transition wouldn't work because it was transitioning background-color, but that wasn't actually changing when the dimmer became active. What actually changed was the opacity. So I changed the transition to track that instead.

Also, I had to make the dimmer element always visible, but with opacity 0 and no pointer events when not active. Otherwise the transition wouldn't work as well, since the visibility would interfere with the other rules.

Finally, I had to change the type of dimmer we were using for the sidebar. The page dimmer is removed from the page when not active, but we need it to be always present so we can have the transition working. Since the dimmer's position was already covering the whole page, I just removed the prop to make it a page dimmer and now we have transition on the sidebar dimmer as expected \o/